Changes default blocklist in the Transmission chart to a working one (#9403)
**Description**

The john.bitsurge.net blocklist (the suggested default in the
Transmission chart) is [no longer hosted at the address mentioned and
never will be
again](https://gist.github.com/johntyree/3331662?permalink_comment_id=3801748#gistcomment-3801748).
I've swapped it out for https://github.com/Naunter/BT_BlockLists/ , a
maintained and seemingly pretty decent alternative? Like John's it also
aggregates a bunch of other lists.

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
I have not deployed this personally to TrueNas Scale yet, although I do
use the blocklist with Transmission on another machine. It seems to
function similarly to John's (except it actually still works). Given
that this is a one-line default variable change to a
non-mission-critical variable it should be good to go? Let me know if it
requires a more rigorous test. :)

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning