feat(authentik): BREAKING CHANGE  Port to new common (#9426)
**Description**
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->
⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [X] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [X] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->
- Installed on native Helm
- Logged in with the bootstrapped password
- Created LDAP and Proxy outposts
- Copied their tokens
- Enabled the outposts in values.yaml and overrode their tokens (this is
necessary to circumvent a known bug
(https://github.com/truecharts/charts/issues/7390) - the bootstrap token
cannot be used by an outpost, so in the future, the overriding of the
token must become mandatory but that's for another PR)
- Logged back in authentik, confirmed the 2 outposts are showing as
healthy in its UI


**📃 Notes:**
<!-- Please enter any other relevant information here -->

- I tried to make the minimal amount of changes necessary to port this
to the new common, in order to keep the diff and this PR as small as
possible
- One small exception to the above, this PR also closes #6986 , since
this was basically 1 line change in the persistence dict
- I have bumped the chart's major revision but i am keeping it in the
incubator train for now for 2 reasons:
* It can get tested by the community (this is my first app port to the
new common)
* There are a few of TODOs left in there (1 by me and several by
others), as well as a couple of known bugs in the chart, which I would
like to address in separate PRs
- I am assuming that `questions.yaml` needs no changes (since I haven't
changed the data structure in `values.yaml`, apart from the necessary
changes to port to new common). I have NOT tested this, so do let me
know if I need to take a look at this and make changes.

**✔️ Checklist:**

- [X] ⚖️ My code follows the style guidelines of this project
- [X] 👀 I have performed a self-review of my own code
- [X] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [X] ⚠️ My changes generate no new warnings
- [X] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [X] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: Kjeld Schouten <kjeld@schouten-lebbing.nl>
Co-authored-by: Kjeld Schouten <kjeld@schouten-lebbing.nl>