feat(common): explicitly set the namespace on all the objects (#488)
**Description**
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->
⚒️ Fixes  #441 

- Added a function to "generate" the namespace and also validate in
terms of length and format.

Current blockers:
Some objects added after the initial refactor use a different code style
and currently is not straightforward to use the above function.

Those will have to be revisited once they have some tests for
themselves.
And after everything else is working and have tests in place (in regards
with the namespace change)

TODO:
  - [x] adjusts regex for validation on name and namespace. 
  - [x] Doc updates

Tests:
  - [x] Configmap
  - [x] CronJob
  - [x] DaemonSet
  - [x] Deployment
  - [x] Endpoint
  - [x] EndpointSlice
  - [x] Job
- [x] NetworkAttachmentDefinition (Due to being only used in Scale and
the way the config is fetched, only global and root level namespaces are
used. You just can't define the NS on the object itself.)
  - [x] PesistentVolumeClaim
  - [x] RBAC
  - [x] Secret
  - [x] Service
  - [x] ServiceAccount
  - [x] StatefulSet

Objects with the implementation (but no tests):
Also those objects do not have per object override and do not support
TPL.
The above issues will be fixed on each one after they are refactored.
  - [x] Certificate
  - [x] Cluster
  - [x] Pooler
  - [x] HorizontalPodAutoscaler
  - [x] Ingress
  - [x] NetworkPolicy
  - [x] PodMonitor
  - [x] PrometheusRule
  - [x] Route
  - [x] ServiceMonitor

**⚙️ Type of change**

- [x] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

**📃 Notes:**
<!-- Please enter any other relevant information here -->

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [x] 📄 I have made corresponding changes to the documentation
- [x] ⚠️ My changes generate no new warnings
- [x] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._