refactor(kitchenowl): resolve issue #8804 (#11337)
**Description**
Kitchenowl was migrated to new common, but deploying ended in "Hmmmm...
couldn't reach server" error and "(111: Connection refused) while
connecting to upstream" log entries.
This PR removes the deprecated "additionalContainer" section and deploys
the backend as a Deployment. Furthermore the secrets template was
refactored to render it instead of hardcoding. Additionally cnpg was
added as it is supported by the upstream project.
As the current chart is not deployable and due to the refactoring of
several parts I was not able to test, but I expect it to be a breaking
change for old deployments as the data is now stored in cnpg instead of
sqlite.

⚒️ Fixes  #8804

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [x] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
Tested on TrueNAS Scale 22.12.3.2 with all operators and best practises
according getting started docs (including Ingress).
After deployment all pods including cnpg are running, no errors in the
logs and application events

**📃 Notes:**

**✔️ Checklist:**

- [ ] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [x] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: polarstack <42521003+polarstack@users.noreply.github.com>